change microseconds to uint64_t

+++ systick.c
@@ -1,9 +1,9 @@
 #include <stdint.h>
 
-uint32_t microseconds = 0;
+uint64_t microseconds = 0;
 uint32_t milliseconds = 0;
 
-uint32_t currentMicroseconds(void)
+uint64_t currentMicroseconds(void)
 {
     return microseconds;
 }
